Migrate BIDS Library to BIDS Common
