+++ tests/test_bidsCommon.py
@@ -1,4 +1,12 @@
-import rtCommon.bidsCommon as bidsCommon
+import logging
+
+from rtCommon.bidsCommon import (
+    getMetadata,
+    loadBidsEntities,
+    metadataFromProtocolName,
+)
+
+logger = logging.getLogger(__name__)
 
 
 def testEntitiesDictGeneration():
@@ -8,7 +16,7 @@ def testEntitiesDictGeneration():
       key: Full entity name, all lowercase
       value: Dictionary with keys "entitity, format, description"
     """
-    entities = bidsCommon.loadBidsEntities()
+    entities = loadBidsEntities()
 
     # Ensure entity count correct
     NUM_ENTITIES = 20  # Manually counted from the file entities.yaml
@@ -28,3 +36,31 @@ def testEntitiesDictGeneration():
     importantKeySample = ["subject", "task", "session"]
     for key in importantKeySample:
         assert key in entities.keys()
+
+
+# Test metadata is correctly extracted from a DICOM to public and private
+# dictionaries by ensuring a sample of public keys have the right value
+def testMetadataExtraction(dicomImage, dicomMetadataSample):
+    public, private = getMetadata(dicomImage)
+    for field, value in dicomMetadataSample.items():
+        assert public.get(field) == str(value)
+
+    # TODO(spolcyn): Also check private keys
+    pass
+
+
+# Test BIDS fields in a DICOM ProtocolName header field are properly parsed
+def testParseProtocolName():
+    # ensure nothing spurious is found in strings without BIDS fields
+    assert metadataFromProtocolName("") == {}
+    assert metadataFromProtocolName("this ain't bids") == {}
+    assert metadataFromProtocolName("nor_is_this") == {}
+    assert metadataFromProtocolName("still-aint_it") == {}
+
+    protocolName = "func_ses-01_task-story_run-01"
+    expectedValues = {'session': '01', 'task': 'story', 'run': '01'}
+
+    parsedValues = metadataFromProtocolName(protocolName)
+
+    for key, expectedValue in expectedValues.items():
+        assert parsedValues[key] == expectedValue
