--- tests/test_bidsLibrary.py
@@ -1,36 +1,0 @@
-import logging
-
-from rtCommon import bidsLibrary as bl
-from rtCommon.bidsCommon import (
-    metadataFromProtocolName,
-)
-
-logger = logging.getLogger(__name__)
-
-
-# Test metadata is correctly extracted from a DICOM to public and private
-# dictionaries by ensuring a sample of public keys have the right value
-def testMetadataExtraction(dicomImage, dicomMetadataSample):
-    public, private = bl.getMetadata(dicomImage)
-    for field, value in dicomMetadataSample.items():
-        assert public.get(field) == str(value)
-
-    # TODO(spolcyn): Also check private keys
-    pass
-
-
-# Test BIDS fields in a DICOM ProtocolName header field are properly parsed
-def testParseProtocolName():
-    # ensure nothing spurious is found in strings without BIDS fields
-    assert metadataFromProtocolName("") == {}
-    assert metadataFromProtocolName("this ain't bids") == {}
-    assert metadataFromProtocolName("nor_is_this") == {}
-    assert metadataFromProtocolName("still-aint_it") == {}
-
-    protocolName = "func_ses-01_task-story_run-01"
-    expectedValues = {'session': '01', 'task': 'story', 'run': '01'}
-
-    parsedValues = metadataFromProtocolName(protocolName)
-
-    for key, expectedValue in expectedValues.items():
-        assert parsedValues[key] == expectedValue
